Issue #117: refactor CredentialsProvider out of DifferentFiles
